Multiple Docker compose files support (#37)
The support for multiple docker compose files has been added in a
backward compatible manner by introducing a new configuration property:

 composeFiles

As name indicates it is a plural form of the original `composeFile`
property which is still present and valid to use.

If the new property `composeFiles` is defined the value of the original
`composeFile` is ignored. A better approach would be to log warning if
user set both the `composeFile` and `composeFiles`. However this is not
possible as the `composeFile` has always a default value, hence is
always defined.